requirements.txt
Flask==1.1.2
Pillow==8.2.0
google-cloud-storage==1.38.0
PyPDF2==1.26.0
reportlab==3.5.67
wand==0.6.6